Flask WTForm vs HTML
Use flask_wtf to vlaidate the user's input because flask_wtf offers
great API's to use and define classes in Python as HTML input elements.
The name of each class-variables is a pointer variable that is a type
of string which yields as an HTML node.

Factorize a file that contains the Login and Register forms. The way to
access the data is by allocating a class and assign that class to a
variable and passes it to an HTML file where the class pointer is
wrraped in an HTML 'form' node.

As a result, this approach is great for accessing user's input.

Note: Its possible to access the requests data by the 'request' context.