User's Posts
Note: I accidantly did not create a new branch beofre adding features.

Provide the ability for the user to submit a post and display on the
screen. This is a nice feature to have as it's the starting block
of this project which is a blog post. Add another form that is a
textarea in which user can writes and submits. Store the user's data
in the session temproraly.

As a result, this a good working feature and we will see how to
customize it on the future.