The beifit of using a database
Use a relational database to connect tables together. In our cause,
create two models Role and User and there are relationship between
those modeles.

Comparing using session as we had, way better than a session in terms
of sapce and long term cached session.

As result, it is good to use a dabase rather than a session