Add change email featrue for the users
Users can be given the option to change their registered email address,
but before the new address is accepted it must be verified with a
confirmation email.

To confirm the email address, a token is emailed to that address.
When the server receives the token back, it can update the user object.