Add roles to the Role table.
The benefit of using powers of two for permission values is that it
allows permissions to be combined, giving each possible combination of
permissions a unique value to store in the role’s permissions field.
For example, for a user role that gives users per‐ mission to follow
other users and comment on posts, the permission value is
FOLLOW + COMMENT = 3.

Overal, this is a very efficient way to store the list of permissions
assigned to each role.